+++ eval.py
@@ -53,7 +53,7 @@ def evaluate(model, testloader, num_classes: int, device):
         masks.mul_(255).squeeze_(dim=1)
 
         # 이미지와 정답 정보를 GPU로 복사
-        images, masks = images.to(device), masks.to(device, dtype=torch.uint8)
+        images, masks = images.to(device), masks.to(device, dtype=torch.int64)
 
         # 예측
         with torch.no_grad():
